Merge and stack text from two columns in `gt` and color one with team colors

The `gt_merge_stack_team_color()` function takes an existing `gt` table and merges column 1 and column 2, stacking column 1's text on top of column 2's. Top text is in all caps while the lower text is bigger, bolded, and colored by the team name in another column. This is a slightly modified version of [`gtExtras::gt_merge_stack()`](https://jthomasmock.github.io/gtExtras/reference/gt_merge_stack.html) written by Tom Mock.

Usage

gt_merge_stack_team_color(
  gt_object,
  col1,
  col2,
  team_col,
  font_size_top = 12,
  font_size_bottom = 14,
  color = "black"
)

Arguments

gt_object

An existing gt table object of class `gt_tbl`

col1

The column to stack on top. Will be converted to all caps, with black and bold text.

col2

The column to merge and place below. Will be smaller and the team color that corresponds to `team_col`.

team_col

The column of team names that match `valid_team_names()` for the color of the bottom.

font_size_top

the font size for the top text.

font_size_bottom

the font size for the bottom text.

color

The color for the top text.

Value

An object of class `gt_tbl`.

Examples

library(gt)
library(mlbplotR)

mlbplotR::load_mlb_teams() %>%
  dplyr::slice(1:5) %>%
  dplyr::select(team_abbr, team_name) %>%
  tidyr::separate(team_name, c("Team1", "Team2","Team3"), fill = "left") %>%
  dplyr::mutate(Team3 = dplyr::if_else(Team3 %in% c("Sox", "Jays"), paste(Team2, Team3), Team3),
                Team2 = dplyr::case_when(grepl("Sox|Jays", Team3) ~ Team1,
                                         Team3 == "Cardinals" ~ paste(Team1, Team2, sep = ". "),
                                         is.na(Team1) ~ Team2,
                                         TRUE ~ paste(Team1, Team2))) %>%
  dplyr::select(team_abbr, Team2, Team3) %>%
  gt::gt() %>%
  gt_merge_stack_team_color(col1 = "Team2", col2 = "Team3", team_col = "team_abbr")
#> <div id="ubbkydxnqq" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ubbkydxnqq .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ubbkydxnqq .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ubbkydxnqq .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ubbkydxnqq .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ubbkydxnqq .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ubbkydxnqq .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ubbkydxnqq .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ubbkydxnqq .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ubbkydxnqq .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ubbkydxnqq .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ubbkydxnqq .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ubbkydxnqq .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ubbkydxnqq .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ubbkydxnqq .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #ubbkydxnqq .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #ubbkydxnqq .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ubbkydxnqq .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #ubbkydxnqq .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ubbkydxnqq .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ubbkydxnqq .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ubbkydxnqq .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ubbkydxnqq .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ubbkydxnqq .gt_left {
#>   text-align: left;
#> }
#> 
#> #ubbkydxnqq .gt_center {
#>   text-align: center;
#> }
#> 
#> #ubbkydxnqq .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ubbkydxnqq .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ubbkydxnqq .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ubbkydxnqq .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ubbkydxnqq .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ubbkydxnqq .gt_two_val_uncert {
#>   display: inline-block;
#>   line-height: 1em;
#>   text-align: right;
#>   font-size: 60%;
#>   vertical-align: -0.25em;
#>   margin-left: 0.1em;
#> }
#> 
#> #ubbkydxnqq .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #ubbkydxnqq .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #ubbkydxnqq .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #ubbkydxnqq .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #ubbkydxnqq .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">team_abbr</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">Team2</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left">ARI</td>
#> <td class="gt_row gt_left"><div style='line-height:10px'>
#> <span style='font-variant:small-caps;font-weight:lighter;color:black;font-size:12px'>Arizona
#> </div>
#> <div style='line-height:12px'>
#> <span style ='font-variant:small-caps;font-weight:bold;color:#A71930;font-size:14px'>Diamondbacks
#> </span></div></td></tr>
#>     <tr><td class="gt_row gt_left">ATL</td>
#> <td class="gt_row gt_left"><div style='line-height:10px'>
#> <span style='font-variant:small-caps;font-weight:lighter;color:black;font-size:12px'>Atlanta
#> </div>
#> <div style='line-height:12px'>
#> <span style ='font-variant:small-caps;font-weight:bold;color:#CE1141;font-size:14px'>Braves
#> </span></div></td></tr>
#>     <tr><td class="gt_row gt_left">BAL</td>
#> <td class="gt_row gt_left"><div style='line-height:10px'>
#> <span style='font-variant:small-caps;font-weight:lighter;color:black;font-size:12px'>Baltimore
#> </div>
#> <div style='line-height:12px'>
#> <span style ='font-variant:small-caps;font-weight:bold;color:#DF4601;font-size:14px'>Orioles
#> </span></div></td></tr>
#>     <tr><td class="gt_row gt_left">BOS</td>
#> <td class="gt_row gt_left"><div style='line-height:10px'>
#> <span style='font-variant:small-caps;font-weight:lighter;color:black;font-size:12px'>Boston
#> </div>
#> <div style='line-height:12px'>
#> <span style ='font-variant:small-caps;font-weight:bold;color:#BD3039;font-size:14px'>Red Sox
#> </span></div></td></tr>
#>     <tr><td class="gt_row gt_left">CHC</td>
#> <td class="gt_row gt_left"><div style='line-height:10px'>
#> <span style='font-variant:small-caps;font-weight:lighter;color:black;font-size:12px'>Chicago
#> </div>
#> <div style='line-height:12px'>
#> <span style ='font-variant:small-caps;font-weight:bold;color:#0E3386;font-size:14px'>Cubs
#> </span></div></td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>