mlbplotR: Create 'ggplot2' and 'gt' Visuals with Major League Baseball Logos

Tools to help visualize Major League Baseball analysis in 'ggplot2' and 'gt'. You provide team/player information and 'mlbplotR' will transform that information into team colors, logos, or player headshots for graphics.

Usage

GeomMLBdarkcaplogo

GeomMLBlightcaplogo

GeomFromPath

GeomRefLines

GeomMiLBlogo

GeomMiLBlightcaplogo

GeomMiLBdotlogo

GeomMLBheads

GeomMLBdotheads

GeomMiLBdotheads

GeomMLBlogo

GeomMLBscoreboardlogo

GeomMLBdotlogo

Format

An object of class GeomMLBdarkcaplogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBlightcaplogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomFromPath (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomRefLines (inherits from Geom, ggproto, gg) of length 6.

An object of class GeomMiLBlogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMiLBlightcaplogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMiLBdotlogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBheads (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBdotheads (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMiLBdotheads (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBlogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBscoreboardlogo (inherits from Geom, ggproto, gg) of length 5.

An object of class GeomMLBdotlogo (inherits from Geom, ggproto, gg) of length 5.

See also

Author

Maintainer: Camden Kay camden.kay23@gmail.com [copyright holder]

Authors:

Sebastian Carl mrcaseb@gmail.com